% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoScan.R
\name{addReferencelabels}
\alias{addReferencelabels}
\title{Add additional annotations for reference data}
\usage{
addReferencelabels(CS, labels)
}
\arguments{
\item{CS}{CytoScan object}

\item{labels}{Vector of numeric or string labels}
}
\value{
CytoScan object
}
\description{
Add additional annotations for reference data
}
\seealso{
\code{\link{addTestlabels}}
}
