% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoScan.R
\name{addTestdata}
\alias{addTestdata}
\title{Add test data to CytoScan object}
\usage{
addTestdata(CS, input, read = FALSE, reload = FALSE, aggSize = 10000)
}
\arguments{
\item{CS}{CytoScan object}

\item{input}{List of FCS file paths}

\item{read}{Whether load subsampled data already (default = FALSE)}

\item{reload}{Reload all data (default = FALSE)}

\item{aggSize}{Size of aggregate sample (default = 10000 cells)}
}
\value{
CytoScan object
}
\description{
Add test data to CytoScan object
}
\seealso{
\code{\link{addReferencedata}}
}
