% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoScan.R
\name{addTestlabels}
\alias{addTestlabels}
\title{Add additional annotations for test data}
\usage{
addTestlabels(CS, labels)
}
\arguments{
\item{CS}{CytoScan object}

\item{labels}{Vector of numeric or string labels}
}
\value{
CytoScan object
}
\description{
Add additional annotations for test data
}
\seealso{
\code{\link{addReferencelabels}}
}
