% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoScan.R
\name{CytoScan}
\alias{CytoScan}
\title{Initialize CytoScan object}
\usage{
CytoScan()
}
\value{
Default initialized CytoScan object
}
\description{
Function to initialize CytoScan workflow. Returns CytoScan object
}
