% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoScan.R
\name{CytoScan_spectral}
\alias{CytoScan_spectral}
\title{Initialize CytoScan object for spectral data}
\usage{
CytoScan_spectral()
}
\value{
Default initialized CytoScan object
}
\description{
Function to initialize CytoScan workflow. Returns CytoScan object
}
