% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoScan.R
\name{Flag}
\alias{Flag}
\title{Flag anomalies based on generated features}
\usage{
Flag(
  CS,
  featMethod,
  flagMethod,
  outlier_threshold = 0.5,
  novelty_threshold = 0.05
)
}
\arguments{
\item{CS}{CytoScan object}

\item{featMethod}{Which features to use for anomaly detection}

\item{flagMethod}{Which type of anomaly detection to use ("outlier" or "novelty")}

\item{outlier_threshold}{Isolation score threshold for outlier detection.}

\item{novelty_threshold}{GMM quantile threshold for novelty detection.}
}
\value{
CytoScan object
}
\description{
Flag anomalies based on generated features
}
\seealso{
\code{\link{generateFeatures}}
}
