% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDemo.R
\name{generateDemo}
\alias{generateDemo}
\title{Generate a folder with a simulated cytometry cohort}
\usage{
generateDemo(dir, nFiles, nCells, nChannels)
}
\arguments{
\item{dir}{Path of directory where FCS files should be saved}

\item{nFiles}{Number of FCS files to generate}

\item{nCells}{Number of cells to simulate per FCS file}

\item{nChannels}{Number of channels to generate per FCS file}
}
\description{
Generate a folder with a simulated cytometry cohort
}
