% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Plot Boxplot}
\usage{
plotBoxplot(
  CS,
  plotData,
  channel,
  n = 1000,
  color = NULL,
  featMethod = NULL,
  nFiles = NULL
)
}
\arguments{
\item{CS}{CytoScan object}

\item{plotData}{Which data to plot based on PCA fit ("test", "reference" or "all")}

\item{n}{Number of cells to sample from each file}

\item{color}{Category used for color ("input", "outlier", "novelty")}

\item{featMethod}{Which features to use when plotting anomalies}

\item{nFiles}{Number of files to plot}
}
\value{
Boxplot
}
\description{
Plot Boxplot
}
