% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(CS, featMethod, plotData = "test")
}
\arguments{
\item{CS}{CytoScan object}

\item{featMethod}{Which features to use}

\item{plotData}{Which data to plot ("test", "reference" or "all")}
}
\value{
Heatmap
}
\description{
Plot heatmap
}
