% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(
  CS,
  featMethod,
  fitData,
  plotData,
  PCx = 1,
  PCy = 2,
  color = NULL,
  shape = NULL,
  plotBars = FALSE,
  plotArrows = FALSE,
  nLoadings = NULL
)
}
\arguments{
\item{CS}{CytoScan object}

\item{featMethod}{Which features to use}

\item{fitData}{Which data to fit PCA on ("test", "reference" or "all")}

\item{plotData}{Which data to plot based on PCA fit ("test", "reference" or "all")}

\item{PCx}{Which PC to plot along x-axis}

\item{PCy}{Which PC to plot along y-axis}

\item{color}{Category used for point color ("outlier", "novelty", "labels")}

\item{shape}{Category used for point shape ("outlier", "novelty", "labels")}

\item{plotBars}{Whether to plot loading coefficients as bars along}

\item{plotArrows}{Whether to plot loadings as arrows}

\item{nLoadings}{Number of loadings to plot. Selects most influential}
}
\value{
PCA plot
}
\description{
Plot PCA
}
