% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoScan.R
\name{processInputCyTOF}
\alias{processInputCyTOF}
\title{Custom preprocessing for CyTOF data}
\usage{
processInputCyTOF(ff)
}
\arguments{
\item{ff}{A flowFrame object}
}
\value{
Transformed flowFrame
}
\description{
Applies arcsinh transformation (cofactor = 5) to biological channels only.
No compensation or unmixing is performed.
}
