% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoScan.R
\name{processInputSpectral}
\alias{processInputSpectral}
\title{Custom preprocessing for spectral flow data}
\usage{
processInputSpectral(ff)
}
\arguments{
\item{ff}{A flowFrame object}
}
\value{
Transformed flowFrame
}
\description{
Applies arcsinh transformation to all channels. No compensation is done.
}
